+++ apps/docs/src/components/cameras/GodCamera.tsx
@@ -1,5 +1,6 @@
 import { PerspectiveCamera } from "@react-three/drei";
-import { useFrame } from "@react-three/fiber";
+import { useFrame, useThree } from "@react-three/fiber";
+import { usePlanet } from "@site/../../packages/react/dist/esm";
 import * as React from "react";
 import { MathUtils, Mesh, Vector2, Vector3 } from "three";
 import { useControls } from "../../hooks/useControls";
@@ -7,45 +8,42 @@ import { useControls } from "../../hooks/useControls";
 const DEFAULT_SENSITIVITY = new Vector2(1, 0.8);
 const DEFAULT_OFFSET = new Vector3(0, 1, 0);
 
-export const FirstPersonCameraSystem: React.FC<{
+export const GodCamera: React.FC<{
   offset?: Vector3;
   sensitivity?: Vector2;
-  position?: Vector3;
-}> = ({ offset = DEFAULT_OFFSET, sensitivity = DEFAULT_SENSITIVITY, position = new Vector3() }) => {
+}> = ({ offset = DEFAULT_OFFSET, sensitivity = DEFAULT_SENSITIVITY }) => {
+
+  const [_position, setPosition] = React.useState(new Vector3())
+  const planet = usePlanet();
+  const { camera } = useThree();
   const maxPitchAngle = 89;
   const minPitchAngle = -89;
   const pitchObjectRef = React.useRef<Mesh>(null);
   const yawObjectRef = React.useRef<Mesh>(null);
   const controls = useControls();
 
+  React.useEffect(() => {
+    // TODO this should be ACTUAL HEIGHT + a value
+    setPosition(
+    _position.set(
+        planet.planetProps.radius * 1.005,
+        0,
+        planet.planetProps.radius * 1.005
+    ).clone());
+    console.log(_position)
+  }, [planet.planetProps.radius]);
+
   useFrame(() => {
     yawObjectRef.current?.position.copy(offset);
     if (pitchObjectRef.current && yawObjectRef.current) {
       const pitchObject = pitchObjectRef.current;
       const yawObject = yawObjectRef.current;
-      // const targetPos = () => {
-      //   return new Vector3(position.x, position.y, position.z);
-      // };
-      // const target = targetPos();
 
       const update = () => {
-        // yawObject.position.x = MathUtils.lerp(
-        //   yawObject.position.x,
-        //   target.x,
-        //   0.1
-        // );
-        // yawObject.position.y = MathUtils.lerp(
-        //   yawObject.position.y,
-        //   target.y,
-        //   0.1
-        // );
-        // yawObject.position.z = MathUtils.lerp(
-        //   yawObject.position.z,
-        //   target.z,
-        //   0.1
-        // );
+        const getLook = () => controls.mouse.move();
+        const getZoom = () => controls.mouse.scroll();
 
-        const getLook = () => controls.mouse.query();
+        console.log(getZoom());
 
         if (controls.mouse.isPointerLocked()) {
           const { x, y } = getLook();
@@ -70,7 +68,7 @@ export const FirstPersonCameraSystem: React.FC<{
   });
 
   return (
-    <group position={position || new Vector3()}>
+    <group position={_position}>
       <mesh ref={yawObjectRef}>
         <mesh ref={pitchObjectRef}>
           <PerspectiveCamera