+++ apps/docs/src/pages/worldbuilder.tsx
@@ -1,7 +1,8 @@
 import BrowserOnly from "@docusaurus/BrowserOnly";
 import * as React from "react";
+
 const LazyWorldbuilder = React.lazy(() =>
-  import("../components/worldbuilder/Worldbuilder")
+  import("../components/world-builder/WorldBuilder")
 );
 
 export default function() {