+++ apps/docs/src/state/store.ts
@@ -1,0 +1,6 @@
+import { makeStore } from "statery"
+import { Mesh } from "three"
+
+export const store = makeStore({
+  sun: null as Mesh | null
+})