+++ packages/planets/src/chunk/createThreadedPlanetWorker.ts
@@ -13,12 +13,18 @@ export function createThreadedPlanetWorker<T, I>({
   colorGenerator: ChunkGenerator3Initializer<T, Color, I>;
 }) {
   let builder: ReturnType<typeof buildChunk<T, I>>;
+  let id: string | null;
   self.onmessage = (msg) => {
     if (msg.data.subject == ChunkBuilderThreadedMessageTypes.INITIAL_DATA) {
       builder = buildChunk<T, I>({
         heightGenerator: heightGenerator(msg.data.initialData),
         colorGenerator: colorGenerator(msg.data.initialData),
       });
+      id = msg.data.id;
+      return;
+    }
+    if (!msg.data.id || msg.data.id !== id) {
+      return;
     }
     if (msg.data.subject == ChunkBuilderThreadedMessageTypes.BUILD_CHUNK) {
       if (!builder) {